test: archive legacy cleanup gates
